gcloudrun/requirements.txt
functions-framework==3.*
google-cloud-firestore
flask==2.0.1
Werkzeug==2.0.3

gunicorn==20.1.0
google-auth==2.3.3
cryptography==36.0.1
solana==0.23.0

